+++ src/Tidier.jl
@@ -6,19 +6,22 @@ using Chain
 using Statistics
 using Reexport
 
+@reexport using DataFrames: DataFrame, describe, nrow, proprow
 @reexport using Chain
 @reexport using Statistics
 
-export Tidier_set, across, desc, starts_with, ends_with, matches, @select, @transmute, @rename, @mutate, @summarize, @summarise, @filter, @group_by, @ungroup, @slice, @arrange, @pull
-
-include("docstrings.jl")
-include("parsing.jl")
+export Tidier_set, across, desc, starts_with, ends_with, matches, @select, @transmute, @rename, @mutate, @summarize, @summarise, @filter, @group_by, @ungroup, @slice, @arrange, @pull, @left_join, @right_join, @inner_join, @full_join
 
 # Package global variables
 const code = Ref{Bool}(false) # output DataFrames.jl code?
 const log = Ref{Bool}(false) # output tidylog output? (not yet implemented)
 
-# Functions to set global variables
+# Includes
+include("docstrings.jl")
+include("parsing.jl")
+include("joins.jl")
+
+# Function to set global variables
 """
 $docstring_Tidier_set
 """